backend-django/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.4
